eatpoint/requirements.txt
Django==4.2.5
autoflake==2.2.1
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.4.0
python-dotenv==1.0.0
django-filter==23.3
djoser==2.2.0
drf-extra-fields==3.7.0
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
django-cors-headers==4.2.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.22
Babel==2.13.0
Pillow==10.0.1
PyJWT==2.8.0
django-cors-headers==4.2.0
python-telegram-bot==20.6
geopy==2.4.0
celery[redis]==5.3.5
django-celery-beat==2.5.0
django-redis==5.4.0
flower==2.0.1